data-collection/requirements.txt
requests>=2.31.0
google-api-python-client>=2.114.0
pandas>=2.1.0
python-dotenv>=1.0.0
fredapi>=0.5.0
